requirements.txt
cx-Oracle==7.1.3
mysqlclient==1.4.2.post1
python-dotenv==0.10.3
virtualenv==16.6.0
